Added basic functionality
Added RulesUtility, to read the rules.cf file and store it in memory.
Added work engines, these will be responsible to make the program work.
Linked the gui and the engines.
Tweaked interface and how it works, to work more efficiently with the
engines.
Created smart hierarchy between classes.
Made interface observer, to link the engine and the respective
interface.